+++ lib/record_validators/command_requests/edit_image.rb
@@ -17,11 +17,23 @@ module RecordValidators
       attr_reader :command_request, :message_text, :picture_id
 
       def valid_message_type?
-        message_text.present? && picture_id.blank? && command_request.prompt.blank?
+        prompt_message? && no_prompt_saved_yet? && source_image_received?
       end
 
       def source_image_ready?
         command_request.latest_image_message&.resolved_image_url.present?
+      end
+
+      def prompt_message?
+        message_text.present? && picture_id.blank?
+      end
+
+      def no_prompt_saved_yet?
+        command_request.prompt.blank?
+      end
+
+      def source_image_received?
+        command_request.latest_image_message.present?
       end
     end
   end
